Add common functionality to the lib
